Updated getUserStatus
Now it sends back the event object instead of just its objectId.  Also
fixed bug where the event query checked if the User was contained in an
array instead of the User’s objectId.  Finally, modified possible
responses to be “FREE” (not part of any events, no invitation),
“ATTENDING” (going to a finalized event), “INVITED” (invited, but
hasn’t RSVPed, to an event), and “WAITING” (part of an unfinalized
event).